third_party/WebKit/LayoutTests/custom-elements/spec/define-builtin-element-expected.txt

This is a testharness.js-based test.
FAIL Element interface for extends is not valid custom element name assert_throws: having valid custon element name element interface (a-a) for extends should throw a NotSupportedError function "() => {
    try {
      func.call(this);
    } catch(e) {
      exception = e;
      throw e;
    }
  }" did not throw
FAIL Element interface for extends defined in specification assert_throws: having element interface for extends (bgsound) undefined in specs should throw a NotSupportedError function "() => {
    try {
      func.call(this);
    } catch(e) {
      exception = e;
      throw e;
    }
  }" did not throw
FAIL localName set to element interface for extends promise_test: Unhandled rejection with value: object "TypeError: Illegal constructor"
Harness: the test ran to completion.
